requirements.txt
PyYAML==6.0
PyYAML==6.0.1
PyYAML==6.0.2
PyYAML==6.0.2
streamlit==1.39.0
streamlit_authenticator==0.4.1
streamlit_authenticator==0.3.2
~treamlit==1.36.0
